% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{generate_probability_question}
\alias{generate_probability_question}
\title{Internal: generate a basic probability question}
\usage{
generate_probability_question(difficulty = "easy")
}
\description{
Internal: generate a basic probability question
}
\keyword{internal}
