% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_one_sample.R
\name{generate_ttest_one_sample_question}
\alias{generate_ttest_one_sample_question}
\title{Internal: generate a one-sample t-test question}
\usage{
generate_ttest_one_sample_question(
  difficulty = "medium",
  n = 30,
  mu0 = 0,
  alpha = 0.05,
  alternative = "two.sided"
)
}
\description{
Internal: generate a one-sample t-test question
}
\keyword{internal}
