% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_two_sample.R
\name{generate_ttest_two_sample_question}
\alias{generate_ttest_two_sample_question}
\title{Internal: generate a two-sample t-test question}
\usage{
generate_ttest_two_sample_question(
  difficulty = "medium",
  n_per_group = 30,
  alpha = 0.05,
  alternative = "two.sided",
  var_equal = FALSE
)
}
\description{
Internal: generate a two-sample t-test question
}
\keyword{internal}
