% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_test.R
\name{generate_which_test_question}
\alias{generate_which_test_question}
\title{Internal: generate a "which test" question}
\usage{
generate_which_test_question(difficulty = "medium")
}
\description{
Internal: generate a "which test" question
}
\keyword{internal}
